Extracted createScopedUriToPath helper
